refactor: dir-metrics
