+++ src/app/home.tsx
@@ -3,10 +3,10 @@ import Activities from "@/components/activities";
 import { BadgeNotificationCheck } from "@/components/badge-notification-check";
 import Hero from "@/components/hero";
 import { LevelUpCheck } from "@/components/level-up-check";
-import MetricsWithSuspense from "@/components/metrics/MetricsWithSuspense";
 import FeaturedMissions from "@/components/mission/FeaturedMissions";
 import MissionsByCategory from "@/components/mission/MissionsByCategory";
 import RankingSection from "@/components/top/ranking-section";
+import MetricsWithSuspense from "@/features/metrics/components/metrics-with-suspense";
 import { generateRootMetadata } from "@/lib/metadata";
 import { checkBadgeNotifications } from "@/lib/services/badgeNotification";
 import { checkLevelUpNotification } from "@/lib/services/levelUpNotification";
