+++ src/features/metrics/components/metrics-error-boundary.tsx
