+++ src/features/metrics/components/metrics-index.test.tsx
@@ -2,13 +2,13 @@ import { render, screen, waitFor } from "@testing-library/react";
 import React from "react";
 
 // メトリクスサービスのモック
-jest.mock("@/lib/services/metrics", () => ({
+jest.mock("../api/metrics", () => ({
   fetchAllMetricsData: jest.fn(),
 }));
 
 import { EXTERNAL_LINKS } from "@/lib/links";
-import { fetchAllMetricsData } from "@/lib/services/metrics";
-import Metrics from "./index";
+import { fetchAllMetricsData } from "../api/metrics";
+import { Metrics } from "./metrics-index";
 
 // モック関数の型アサーション
 const mockFetchAllMetricsData = fetchAllMetricsData as jest.MockedFunction<
